updates to server implementation
Added documentation to undocumented code in client.java and server.py.
ensured client.java receives an appropriate workout calendar json implementation that can be parsed and applied to the user logged in after the server generates it.
Worked on the workout generator aspect of the server and ensured it was generating workouts correctly.