Intensity and workout generator changes
Edited intensity to be a threshold to add exercises up to. Added an intensity value to each individual exercise in the exercise data. Added repetition information to each exercise description. Edited workout Generator to generate workouts up to the intensity threshold. Edited workout to have methods that derive sets based on duplicate workouts. Edited daily details controller to display sets of the exercise that the user should do.